1. Fixed issue where a change in spellId for spells that have two different versions (and spellIDs) based on the current stance wasn't detected during a stance change. This causes issues when determining the key binding.
Cause: The registered events in ActionBar.lua happen before the spellID changes during a stance changes.
Fix: Added a timer that fires one second after a stance change to rescan the spells.

2. Fixed issue where key binding for spells that have a cooldown and are on both active slots and on inactive bonus bars were incorrect.

Cause: When a spell with a cooldown is activated, blizzard fires an ACTIONBAR_SLOT_CHANGED event for every slot the spell is present in, even if that slot is on an inactive bonus bar. Unlike the code for UpdateActionSlots which only calls UpdateActionSlot for active slots, the event handler for ACTIONBAR_SLOT_CHANGED calls UpdateActionSlot for every slot change event, even if it is for an inactive slot. This can cause a keybinding to be incorrect.
Fix: Added code to ACTIONBAR_SLOT_CHANGED event handler to prevent calling UpdateActionSlot for inactive slots.