Enhance ItemRequire and SpellRequire to allow "OR" conditionals.
ItemRequire() and SpellRequire() had a limitation where a property
was set to a particular value by associating only a single
requirement handler for that value.  For example:

   SpellRequire(stealth unusable 1=stealthed,1)
   SpellRequire(stealth unusable 1=combat,1)

Only the second handler "combat=1" is used because only one handler
could be associated with "unusable=1" (the second SpellRequire
effectively overwrote the first).

This commit allows for handlers for multiple requirements to be
collected into an array associated with that property's value.
Checking requirements now means checking if any of the handlers in
the array are true, effectively turning multiple ItemRequire() or
SpellRequire() commands into "OR" conditionals.

Fixes issue #530 where rogues are recommended to Stealth even while
stealthed.