Add the correct spell ID for the Shadow Dance buff to spell lists.
The Stealthed() condition just returns the presence of any of buffs
listed in "stealthed_buff". Add the correct spell ID for the Shadow
Dance buff to the list, which differs from the spell ID of the
spell itself. Also make the same change for "rogue_stealthed_buff"
for completeness.

This addresses issue #714.