Teach the Covenant class to manage its own ID <-> name mapping.
Enumerate all of the covenents and their corresponding IDs in a
COVENENT_ID_BY_NAME table and allow using names with the
IsCovenant() condition.

Modify the IsCovenant() condition to accept either a number or a
string for the parameter name.

This removes the need for the defintions in scripts/ovale_common
for `kyrian', `venthyr', `night_fae', and `necrolord', as these
strings can be used directly by IsCovenant().