Remove bitrotted and likely incorrect code regarding "nocd".
At one point, it may have been set as a paramter to AddIcon to
ignore the cooldown of the spell being displayed, but that hasn't
been used in at least 12 years, and the usage of "nocd" somehow
changed over time without any documentation on the new semantics.

Removing it to make the remaining codepath slightly clearer.