fix: register correct callback to AceTimer:ScheduleTimer()
AceTimer:ScheduleTimer() expects "PrintVersionCheck" to be a method
of "this.module", but no such method exists.

Use the correct function "this.PrintVersionCheck" and pass "this"
as a parameter to the function so that the method is called on the
correct object.

This fixes "/ovale ping" to allow checking for other users of Ovale
in the group or in the guild (if not in a group).