fix: 'extraPower' parameter of TimeToPower() not used correctly
The commit 18d77148dfd32c04b7b2c9f559b0aca624edb8c2 did not
correctly handle the optional extraPower paramter of
Power:TimeToPower().

Generalize the parameter to be table whose keys "extra_energy" and
"extra_focus" are checked for extra amounts of the power type when
calculating the number of seconds until that power threshold is
met.

This removes the need to infer the pooled resource/power type.