fix: pet SimC actions for deathknights
Blood and frost death knights summon a ghoul with Raise Dead that
the game treats as a totem, so use TotemPresent() and
TotemRemaining() to translate "active" and "remains".

Unholy death knights summon a true pet ghoul, but also have three
abilities that summon minions that are not true pets: Army of the
Dead, Apocalypse, and Summon Gargoyle. Army of the Dead and
Apocalypse ghouls are neither pets nor totems, so translate
pet.*.active by comparing the remaining spell cooldown with the
cooldown duration. Summon Gargoyle functions as a totem, so use
TotemPresent() and TotemRemaining() to translate "active" and
"remains".