feat(simulationcraft): improve meaning of "desired_targets"
In SimulationCraft, "desired_targets" is the number of long-lived
enemies that exist for the duration of the fight, and is less than
or equal to "active_enemies", which is the total number of enemies
present. Instead of translating "desired_targets" into
"Enemies(tagged=1)", which is the number of active enemies that the
player has attacked, change it to a custom function that by default
returns 1, but can be modified by a dropdown menu option.

Currently, support 1, 2, or 3 desired targets.

Fixes issue #236.