fix(runner): remove erroneous "constant" optimization
The "constant" optimization in RecursiveCompute() was returning the
wrong results for "if" nodes when the result was a "value" node
because the cached result from a previous computation wasn't being
updated if the condition's computed result was different than
before.

Remove the "constant" optimization from RecursiveCompute() and add
a new visitor computeUndefined() to visit any nodes of type
"undefined" that were generated by parsing typed functions.

Export type AstUndefinedNode in engine/ast so that in can be used
in engine/runner in computeUndefined().

This fixes part of issue #749.