fix(best-action): always use the correct texture for actions
The idiom that was used to set the texture was unnecessary when
the variables were all local, but is plain wrong when using
"result" fields because it would never overwrite a cached result's
texture. Simplify by using a simple if-else statement to set the
texture.

This fixes the rest of issue #749.