fix: script translations of totem pets (#768)
* fix: allow the states/Totem module to support all classes

Every spell or item that summons minions has those minions treated
as totems by the game. Their information can be retrieved via the
Blizzard API function GetTotemInfo(slot). Treat every class as a
totem-capable class and track their totem states appropriately
through game events.

* refactor: in EmitOperandPet(), dispatch first to EmitOperandSpecial()

Try to pattern-match for EmitOperandSpecial() before falling
through to the rest of the pet operand logic. This allows for
checking for specific problematic actions that need to be handled
specially before doing the usual pet operand logic.

Verified to emit the same scripts as before this change.

* fix: pet SimC actions for deathknights

Blood and frost death knights summon a ghoul with Raise Dead that
the game treats as a totem, so use TotemPresent() and
TotemRemaining() to translate "active" and "remains".

Unholy death knights summon a true pet ghoul, but also have three
abilities that summon minions that are not true pets: Army of the
Dead, Apocalypse, and Summon Gargoyle. Army of the Dead and
Apocalypse ghouls are neither pets nor totems, so translate
pet.*.active by comparing the remaining spell cooldown with the
cooldown duration. Summon Gargoyle functions as a totem, so use
TotemPresent() and TotemRemaining() to translate "active" and
"remains".

* fix: correct the list of totem spells for monks

Remove Niuzao and Xuen, which are actual pets. Add Chi-Ji and Yulon
for mistweavers which are totems.