fix: check for correct buffs on pets (#786)
Dark Transformation (unholy DK) and Beast Cleave (BM hunter) are
buffs that need to be checked on the pet, not the player. Fix the
names of the buffs that are checked within EmitOperandBuff() that
trigger redirecting the target from "target" to "pet".

Use the correct spell ID for the pet Frenzy buff applied by Barbed
Shot and check for the buff on the pet, not on the player.