fix(aura): get combopoint state for computing aura duration
GetBaseDuration() is meant to return the duration of an aura if it
was applied by a specific spellcast or in the current state. When
called with no spellcast argument, it was assuming zero combopoints
and not correctly extending the duration via the property
"add_duration_combopoints". In this case, grab the number of
combopoints from the current state instead.

Since GetBaseDuration() now uses state, adjust it to take atTime
as an optional parameter to figure out which state to use.