fix(druid): many fixes and improvements to feral script (#803)
Prowl is only usable out-of-combat and not prowling

Add damage multipliers for DoT snapshotting

The following buffs work with feral snapshotting:

* Tiger's Fury buffs Rake, Rip, Thrash, and Moonfire.
* Bloodtalons buffs Rip.
* Improved Rake (from stealth) buffs Rake.

Add the corresponding SpellDamageBuff() lines to the spell
definitions so that the *Multiplier() conditions will return the
correct values instead of the default of 1.

Restrict certain spells to specs or affinities

Balance specialization or Balance Affinity is required for casting
Moonkin Form, Starfire, Starsurge, Sunfire, or Typhoon.

Guardian specialization or Guardian Affinity is required for
casting Frenzied Regeneration or Incapacitating Roar.

Feral specialization or Feral Affinity is required for Maim, Rake,
Rip, Main, or Swipe.

moonfire_cat is only usable with Lunar Inspiration

* fix(aura): get combopoint state for computing aura duration

GetBaseDuration() is meant to return the duration of an aura if it
was applied by a specific spellcast or in the current state. When
called with no spellcast argument, it was assuming zero combopoints
and not correctly extending the duration via the property
"add_duration_combopoints". In this case, grab the number of
combopoints from the current state instead.

Since GetBaseDuration() now uses state, adjust it to take atTime
as an optional parameter to figure out which state to use.

* fix(druid): duration of Rip depends on the number of combopoints

* fix(conditions): return the correct value for TickTimeRemaining()

The return value of TickTimeRemaining() isn't a constant, since it
varies over a time interval. Use the correct (value, origin, time)
triplet instead.

* feat: support spells applying an aura with different durations

This is extremely basic support for an aura that is applied at two
different durations by two different spells. The spell identified
with "half_duration=spellId" in SpellInfo() causes the base
duration calculation to be halved.

* fix(druid): set Primal Wrath to cause half-duration Rip

* fix(aura): implement ticksGainedOnRefresh()

Add a basic implementation of ticksGainedOnRefresh() to replace the
incorrect "ReturnConstant(0)".

* fix(emiter): translate SimC's druid.<spell>.ticks_gained_on_refresh

Add a new section to EmitOperandSpecial() to parse and translate
"druid.<spell>.ticks_gained_on_refresh", with a special case for
primal_wrath.

* fix(emiter): require more than one enemy for Primal Wrath

The current SimC feral druid profile used as a basis for the
default feral Ovale script sometimes suggest Primal Wrath on single
target. Fix this by emitting an extra condition to check for more
than one enemy when suggesting the Primal Wrath action.

* fix(druid): regenerate feral druid scripts

Regenerate feral druid scripts for ticks_gained_on_refresh and
primal_wrath changes to simulationcraft/emiter.ts.