fix: properly check for known spells in InUsableSpell()
Revert 86439eb31f273ab217bfd4decce845f88b09f7f2 and fix the
underlying problem in a different way. That commit used the WoW
IsUsableSpell() API function to determine whether a spell is known,
but that returns false if the spell is genuinely not usable due to
either cooldown or power requirements. Fix this in a better way by
teaching SpellBook.ts:IsKnownSpell() to check for a spell being
known to the player even if it's not in the spellbook by using the
WoW IsSpellKnown() API function.

While here, adjust IsUsableSpell() so that there is only a single
exit point from the function to make the profiler stats work
correctly.