fix(simulationcraft): cleanup of simulationcraft APL translator (#816)
* fix(simulationcraft): druid's Pulverize requires 2 stacks of Thrash

Move the condition code for Pulverize from an explicit condition in
the script to a spell requirement. Fix the condition code to check
for at least 2 stacks of Thrash instead of the ancient mechanic of
simply needing the Thrash debuff to be on the target.

* fix(simulationcraft): remove check for guardian druid Rejuvenation

Enhanced Rejuvenation for guardian druids was removed from the game
in Legion.

* fix(simulationcraft): feral druid Regrowth no longer procs Bloodtalons

The feral druid Bloodtalons talent was changed in Shadowlands and
no longer procs from casting Regrowth.

* fix(simulationcraft): hunter Kill Command requires an active pet

Kill Command requires that the hunter's pet be present and free to
attack the target. Move the condition from an explicit check in the
script to a spell requirement.

* fix(simulationcraft): remove Arcane Brilliance translation

Arcane Brilliance was removed from the game a while ago, and will no
longer be encountered as "arcane_brilliance" in an APL.

* fix(simulationcraft): frost mage can't summon pet if pet exists

Frost mages can't cast Summon Water Elemental if a pet is already
pre-existing. Move the condition check from an explicit check in
the mage script to a spell requirement.

* fix(simulationcraft): remove nimble_brew translation

Monk's Nimble Brew was removed from the game a while ago, and will
no longer be encountered as "nimble_brew" in an APL.

* fix(simulationcraft): set conditions for Whirling Dragon Punch to be usable

Monk's Whirling Dragon Punch can only be cast when both Fists of
Fury and Rising Sun Kick are on cooldown. Move the condition to
check this from the script into a spell requirement.

* fix(simulationcraft): premeditation is no longer an action

Rogue Premeditation was changed a while back to be a buff you gain
from entering Stealth, not a spell you use out of combat to gain
combo points. Remove the translation for the premeditation action.

* fix(simulationcraft): magma_totem is no longer a valid action

Shaman's Magma Totem was removed from the game a while ago. Remove
the translation for magma_totem.

* fix(simulationcraft): warrior Battle Shout is not available to tanks

* fix(simulationcraft): warrior Commanding Shout is not available

Warrior Commanding Shout was removed from the game a while ago.
Remove the translation of commanding_shout.

* fix(simulationcraft): warrior Enraged Regeneration doesn't give health

Warrior Enraged Regeneration was changed a while ago to only be
usable by Fury warriors and no longer grants health. Remove the
special condition to check for missing health when translating
enraged_regeneration.

* fix(simulationcraft): warrior Heroic Charge is not available

Warrior Heroic Charge was removed from the game a while ago.
Remove the translation for the heroic_charge action.

* fix(simulationcraft): 'stance' action is no longer available

Stance-dancing is not available in SimulationCraft or in-game.
Remove the translation for the 'stance' action.