fix(simulationcraft): ignore storm_earth_and_fire_fixate for monks
The Storm, Earth, and Fire: Fixate spell can be cast off-GCD at any
time, but there is no way to tell if the copies are actually
fixated on a single target or not. Just ignore the spell action so
that the script does not ask the player to keep spamming Fixate.