fix: `BuffComboPoints()` condition is no longer implemented
Auras have not saved combo point information for multiple
expansions, and the script condition has not been used for a
similar amount of time. Change the `BuffComboPoints()` and
`DebuffComboPoints()` script conditions to "not implemented" and
return a constant value of zero.

SimulationCraft also no longer supports looking up the number of
combo points that an aura was applied with, so we can probably
safely remove these two conditions in a future Ovale release.