refactor: remove snapshots (#923)
* refactor: remove references to `snapshotTime`

Initial work on removing snapshots from the codebase. This commit
removes references to `snapshotTime` from all relevant modules as
a prelude to removing snapshots completely.

* refactor: aura stats can't be rolled over when auras are refreshed

Blizzard removed the concept of player stats being snapshot into
auras or being refreshed when an aura is refreshed before it falls
off, so remove it from the Aura module.

* refactor: remove references to `PaperDollSnapshot`

We no longer use snapshots, although we do copy some information
from the current state, e.g., spell-specific damage multiplier,
etc.

Also rearrange a few functions so that `registerSpellcastInfo`,
`copySpellcastInfo`, and `saveSpellcastInfo` are used correctly.

Garbage-collect the code in the Power module to save power info
into a spellcast as it was never actually used.

Rename `Future:updateSpellcastSnapshot()` to `updateSpellcastInfo()`
and re-use the existing `saveSpellcastInfo()` to update a
spellcast's information to the state at a given time.

These changes remove snapshots entirely from the codebase.

* fix: `BuffComboPoints()` condition is no longer implemented

Auras have not saved combo point information for multiple
expansions, and the script condition has not been used for a
similar amount of time. Change the `BuffComboPoints()` and
`DebuffComboPoints()` script conditions to "not implemented" and
return a constant value of zero.

SimulationCraft also no longer supports looking up the number of
combo points that an aura was applied with, so we can probably
safely remove these two conditions in a future Ovale release.

* style: directly use imported module functions instead of aliases

Directly use the following imported functions from the `string`
and `table` modules instead of using aliases:

    `strlower` -> `lower`
    `strsub`   -> `sub`
    `tconcat`  -> `concat`
    `tremove`  -> `remove`