refactor: new soulbind module that supersedes conduit module
Rename the `Conduit` class to `Soulbind` to better reflect the
scope of the information provided. The `Soulbind` class API is a
superset of the `Conduit` class API and provides script conditions
to query for:

  - active conduits by spell ID (`Conduit()`)
  - active soulbind traits by spell ID (`Soulbind()`)

It caches the data to avoid using `C_Soulbinds` functions to query
the server for the information on every refresh of the Ovale frame.

This also fixes the `Soulbind()` condition to check for an active
soulbind **trait** instead of checking for the active soulbind
itself, and now matches how the default scripts are using the
`Soulbind()` condition.