fix: only update weapon damage stats if a weapon slot changed
PLAYER_EQUIPMENT_CHANGED is a bursty event, so the event handlers
for it must handle events quickly.  Ignore PLAYER_EQUIPMENT_CHANGED
if the slot changed isn't a weapon slot since we only listen for
this event to update the weapon damage stats.

This fixes an issue with the game client hanging when changing
specializations and an new equipment set is auto-equipped.