fix: modify for loop to match intentions
The local _repeat function is meant to repeat a string N times, so
adjust the loop limits appropriately to do that.

This is a small functional change where previously due to an error
in codegen by `tstolua`, `_repeat(s, 1)` would return `s + s`
instead of just `s`, even though the TypeScript was correct.
However, the exported function `printRepeat` is only used in an
unused `toString` method of the `SimulationCraft` class, so there
is no functional impact.