feat: add doubly-linked list implementation
Add doubly-linked list implemented as a circular list. The API of
the `List` class is almost identical to the `Deque` API.

A `List` uses more memory than a `Queue`, but a `List` takes O(1)
time to do insertions and removals anywhere in the list, whereas a
`Queue` only takes O(1) time to do insertions and removals at the
front and back of the queue, and O(n) everywhere else.

Also add a testsuite for the `List` class.