feat: add CombatLogEvent module to capture CLEU events
Handling COMBAT_LOG_EVENT_UNFILTERED is messy, with the number of
return values from CombatLogGetCurrentEventInfo() varying based on
the CLEU sub-event, and the the return value types varying based on
the prefix and suffix of the sub-event.

Encapsulate all of the processing in a single class and use a
type-safe method to access the event parameters.

Provide type-safe defaults for all parameters so we don't need
to check for undefined values.

For now, make the `Ovale_CombatLogEvent` message mirror the usual
`COMBAT_LOG_EVENT_UNFILTERED` event, but in the future, we can
reduce the number of dispatches by keeping track of which subevents
are registered by which modules, and instead of hooking
`Ovale_CombatLogEvent`, they could hook the individual subevents
instead. For example, only the modules that watch for
`SPELL_CAST_SUCCESS` would have their callbacks invoked, instead
of every single event handler that's watching for
`Ovale_CombatLogEvent`.