refactor: move COMBAT_LOG_EVENT-related code to a new module (#976)
* feat: add CombatLogEvent module to capture CLEU events

Handling COMBAT_LOG_EVENT_UNFILTERED is messy, with the number of
return values from CombatLogGetCurrentEventInfo() varying based on
the CLEU sub-event, and the the return value types varying based on
the prefix and suffix of the sub-event.

Encapsulate all of the processing in a single class and use a
type-safe method to access the event parameters.

Provide type-safe defaults for all parameters so we don't need
to check for undefined values.

For now, make the `Ovale_CombatLogEvent` message mirror the usual
`COMBAT_LOG_EVENT_UNFILTERED` event, but in the future, we can
reduce the number of dispatches by keeping track of which subevents
are registered by which modules, and instead of hooking
`Ovale_CombatLogEvent`, they could hook the individual subevents
instead. For example, only the modules that watch for
`SPELL_CAST_SUCCESS` would have their callbacks invoked, instead
of every single event handler that's watching for
`Ovale_CombatLogEvent`.

* refactor: make modules use CombatLogEvent instead of CLEU directly

* refactor: adjust DemonHunterDemonic module to use CombatLogEvent

This commit also fixes the module to actually work. Previously, the
CLEU event was registered, but the name of the handler function
wasn't specified, so nothing was actually being run when the event
fired. This commit hooks up the handler function to watch for the
new `Ovale_CombatLogEvent` message.

* refactor: adjust SpellDamage module to use CombatLogEvent

This commit also fixes a bug where periodic damage from DoTs was
not correctly being counted toward the total damage by a spell
because the event being checked was `SPELL_PERIODIC_AURA` instead
of `SPELL_PERIODIC_DAMAGE`.

* refactor: only trigger callbacks for registered CLEU events

Modify `registerEvent()` to take a callback function that is
invoked when the corresponding CLEU event is fired.

Add a `fire()` method to trigger invocation of all of the callbacks
associated with a CLEU event.

Remove the short-lived `Ovale_CombatLogEvent` message.

* refactor: use new API to register for specific CLEU events

* refactor: handle event handlers registering other events

* refactor: improve readability