fix: resolve spells in SpellCooldown() condition before checking
`SpellCooldown()` can take a list of spell IDs and returns the
earliest time that one is available. Resolve the spell IDs to their
final replacement spells if they were replaced because the WoW API
function `GetSpellCooldown()` returns 0 if used with original spell
ID, even though it's been replaced in the spell book.