fix: simplify identifying runeforges by solely using bonus IDs
Add a helper table generated by `importsimc` into the `dbc` module
that contains a mapping of runeforge spell IDs to item bonus IDs.

Use the `runeforgeBonusId` mapping to get the bonus ID for a
runeforge and compare that against the list of bonus IDs for an
item to determine if the item has that runeforge.