fix: rewrite DemonHunterSigils module to properly handle state
Keep track of current sigils that are charging (before activation)
and also keep track of sigils in the simulator that are charging
due to a spell cast within the simulator.

Don't look for an activation spell because if a sigil doesn't hit
anything, then it'll never seem like it activated. Just assume that
it will activate after the charging time has ended.

Also add Elysian Decree to the sigils that are tracked.