fix: rewrite demon hunter modules to correctly handle state (#1015)
* fix: rewrite DemonHunterDemonic module to properly handle state

The `DemonHunterDemonic` module doesn't have to gain or lose fake
auras to correctly track Demonic in Shadowlands. Only add a
Metamorphosis buff of the correct duration within the simulator
if a Demonic trigger spell was cast, for either Havoc or Vengeance
specializations.

* fix: rewrite DemonHunterSigils module to properly handle state

Keep track of current sigils that are charging (before activation)
and also keep track of sigils in the simulator that are charging
due to a spell cast within the simulator.

Don't look for an activation spell because if a sigil doesn't hit
anything, then it'll never seem like it activated. Just assume that
it will activate after the charging time has ended.

Also add Elysian Decree to the sigils that are tracked.

* fix: rewrite DemonHunterSoulFragments module to properly handle state

Keep track of pending Soul Fragments that have not yet spawned and
also keep track of the total number of Soul Fragments after spells
have been cast within the simulator.

* fix: typo for "vengeance"