fix: ensure that pending Soul Fragments can never be negative
Use `math.max` and `math.min` to express limits instead of inlining
computations and add a check to enforce that `pending` can never be
negative.

Refactor the code slightly to extract a function to update
`this.current.count` to be called whenever `count` or `pending` has
changed, instead of duplicating the same code three times, and
enforce limits appropriately in the new function.