fix: support (trivially) rune_word SimC operand
The `rune_word.<shard_type>.<property>` SimC operand checks for
properties of the Shard of Domination bonus if equipped. Assume it
is always true for now since its only use in the APLs is for
subtlety rogue `rune_word.frost.enabled`, but leave a TODO comment
for properly implementing this later.